fix configmap
